+++ packages/auth/src/plugins/plugins.prisma.ts
@@ -1,4 +1,4 @@
 import '@lumina/env'
-import { prisma as database } from '@repo/database'
+import { prisma as database } from '@lumina/db'
 
 export const prisma = database