+++ tooling/prettier-config/index.js
@@ -1,0 +1,7 @@
+/** Shared Prettier config for Lumina workspaces. */
+export default {
+  semi: false,
+  singleQuote: true,
+  trailingComma: 'es5',
+  printWidth: 100,
+}